Science and Safety
A few more headings at the beginning to put safety first. Also, to try
and figure out just what the heck I'm doing with git and the editing
process. So I commit to the develop branch and then submit a pull
request for the merging to the master. Each requires a title and
description